# Vendor JS moved into webpack