--- a/openlibrary/plugins/openlibrary/js/index.js
+++ b/openlibrary/plugins/openlibrary/js/index.js
@@ -1,3 +1,29 @@
+import 'jquery';
+import 'jquery-migrate';
+// Slick#1.6.0 is not on npm
+import '../../../../vendor/js/slick/slick-1.6.0.min.js';
+// npm jquery-ui@1.12.1 package does not match the one we have here, so for now we load from vendor
+import '../../../../vendor/js/jquery-ui/jquery-ui-1.12.1.min.js';
+// For dialog boxes (e.g. add to list)
+import '../../../../vendor/js/colorbox/1.5.14.js';
+// jquery-sparkline @ 1.4.3
+import '../../../../vendor/js/jquery-sparkline/jquery.sparkline.js';
+// jquery-show-password#1.0 not on npm, no longer getting worked on
+import '../../../../vendor/js/jquery-showpassword/jquery.showpassword.js';
+// jquery.form#2.36 not on npm, no longer getting worked on
+import '../../../../vendor/js/jquery-form/jquery.form.js';
+// jquery-validate#1.6 not on npm
+import '../../../../vendor/js/jquery-validate/jquery.validate.js';
+// jquery-autocomplete#1.1 with modified
+import '../../../../vendor/js/jquery-autocomplete/jquery.autocomplete-modified.js';
+// unversioned.
+import '../../../../vendor/js/wmd/jquery.wmd.js'
+// jquery-flot 0.7.0
+import '../../../../vendor/js/flot/jquery.flot.js';
+import '../../../../vendor/js/flot/jquery.flot.selection.js';
+import '../../../../vendor/js/flot/jquery.flot.crosshair.js';
+import '../../../../vendor/js/flot/jquery.flot.stack.js';
+import '../../../../vendor/js/flot/jquery.flot.pie.js';
 import { validateEmail, validatePassword } from './account.js';
 import autocompleteInit from './autocomplete';
 import addNewFieldInit from './add_new_field';
@@ -78,8 +104,10 @@ jQuery.fn.removeHighlight = removeHighlight;
 jQuery.fn.highlight = highlight;
 jQuery.fn.fadeToggle = fadeToggle;
 
+window.jQuery = jQuery;
+window.$ = jQuery;
 // Initialise some things
-$(function () {
+jQuery(function () {
     initValidate($);
     autocompleteInit($);
     addNewFieldInit($);