The script static/js/vendor.jsh previously constructed a vendor.js
bundle that was shipped to users. This meant every page had 2 JS
files to download. This moves all those libraries verbatim, using
npm pinned versions where possible to webpack and imports them all
in the exact same order at the top of index.js

This has 2 benefits:
* We now only load 1 JS file

This also enables me to do 3 follow ups:
* identify jquery migration issues in the console and fix them
* move vendor bundles closer to where they are actually used
* investigate upgrading libraries
* lazy load code on pages where it's not needed

### Description
This is a non-functional change. The way we generate code has changed, not the code itself.

Closes #

### Technical
Hopefully patch is self explanatory, 
The contents of static/js/vendor.jsh have moved to openlibrary/plugins/openlibrary/js/index.js - now webpack generates all our JS.
Several libraries are now loaded from npm pinned versions.

### Testing
- you'll need to run npm install and `make js` to see this in action.
- confirm only one JS file is now loaded per page
- I've checked various pages - subject, homepage, books page, edits page for JS errors and haven't seen any. Note you may notice some warnings in the console from jquery migrate - it's great to finally see these so we can give them our attention! Another pair of eyes doing the same would be useful.

### Evidence
There should be no UI changes with this change.