@cdrini 2019-08-18
Did some perf tests; time is computed using Firefox's throttling for "Regular 3G", averaged over 5 runs, with caching disabled. Tested on page http://192.168.99.100:8080/books/OL2058361M/The_wit_wisdom_of_Mark_Twain

Before: 

| File         | Transferred | Size    | Time  |
| ------------ | --------- | --------- | ----- |
| vendor-v2.js | 575.76 KB | 575.54 KB | 6.7 s |
| all.js       | 133.10 KB | 132.87 KB | 5.0 s |
| Net          | 708.86 KB | 708.39 KB | 6.7 s |

After:
| File         | Transferred | Size    | Time  |
| ------------ | --------- | --------- | ----- |
| all.js       | 669.04 KB | 669.04 KB | 7.1 s |
| Net          | 669.04 KB | 669.04 KB | 7.1 s |

We don't have gzipping enabled locally (unfortunately), so the real time will be less though, so I don't think it will make much of a difference, so perf is ok.