@cdrini 2019-08-18
It looks like the markdown previews are broken now; maybe wmd needs to be loaded in index.js?
![image](https://user-images.githubusercontent.com/6251786/63218842-ce966880-c132-11e9-88dc-551813067514.png)