@jdlrobson 2019-08-18
> It looks like the markdown previews are broken now; maybe wmd needs to be loaded in index.js?
> ![image](https://user-images.githubusercontent.com/6251786/63218842-ce966880-c132-11e9-88dc-551813067514.png)
> ^ http://localhost:8080/works/OL53924W/The_complete_works_of_Mark_Twain/edit

Yeh looks like I didn't copy that over. Will do.

Regarding gzip, bundlesize will tell you this - and we define in package.json . Before vendor-v2.js was 164KB with gzipping and all.js was 35.5KB. The combined gzip now is 201.2kb (I'm a bit surprised as I could have sworn it went down when first worked on this patch but it's only a minor increase, and having one bundle will allow us to code split which on the long term will speed things up considerably - see a2f27fec58d74145677547a4674d5f6e361be1fe and eb97d05b245cd175ebdb561399359d229562cf56 for a sneak peak)