@jdlrobson 2019-08-18
Would it be better to link to the entire docs at the top of the file? I'm
just a bit worried we will have lots of lines pointing to sub pages that
might change URL with future webpack releases.

Agree that config is difficult to grok. We have a dev at wikimedia we call
the webpack guy 😂


On Sat, Aug 17, 2019, 7:47 PM Drini Cami <notifications@github.com> wrote:

> *@cdrini* commented on this pull request.
> ------------------------------
>
> In webpack.config.js
> <https://github.com/internetarchive/openlibrary/pull/2313#discussion_r314966260>
> :
>
> > @@ -78,6 +84,8 @@ module.exports = {
>      devtool: 'source-map',
>
>      performance: {
> +        maxAssetSize: 670 * 1024,
> +        maxEntrypointSize: 670 * 1024,
>
> That's fair; maybe add that link to the performance object? Build config
> files are kind of difficult to grok, so making it easier for future folk to
> understand what's happening makes them less intimidating.
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/internetarchive/openlibrary/pull/2313?email_source=notifications&email_token=AABEKEFSBN27T75JV473DU3QFCZ35A5CNFSM4IMK4Q62YY3PNVWWK3TUL52HS4DFWFIHK3DMKJSXC5LFON2FEZLWNFSXPKTDN5WW2ZLOORPWSZGOCB3V6XQ#discussion_r314966260>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AABEKEE6FA3MUNRZ3MRQQTLQFCZ35ANCNFSM4IMK4Q6Q>
> .
>