@jdlrobson 2019-08-19
> The markdown previews are still not working; are they working for you?
> 
> (Also, those sneak peak commits look glorious :) )

Okay so I debugged this. The jquery.wmd library does this:
```
var Attacklab = Attacklab || {}
```
it expects that to register a global, which is no longer the case, which breaks everything.
To fix this we'll need to add:
```
window.Attacklab = Attacklab
```
at the bottom of jquery.wmd

This can be done with https://github.com/internetarchive/wmd/pull/4 and an update to vendor in the codebase. Will need help from @mekarpeles on this one. Then this patch should just work.