@cdrini 2019-08-19
In the interest of getting this out the door, could we include the `wmd` file in a `<script>` tag directly and remove the JS import? I want to merge this PR :)

(I'm also nervous to edit the wmd code and would rather avoid it if possible :P)