@jdlrobson 2019-08-19
> In the interest of getting this out the door, could we include the `wmd` file in a `<script>` tag directly and remove the JS import? I want to merge this PR and take it off my to-do list :)
> 
> (I'm also nervous to edit the wmd code and would rather avoid it if possible :P)

vendor is not publicly accessible, so to do this we'd need to keep that Vendor make file to copy it across. We'd then have to load it on every page with an additional request (after the core JS file) I'd rather not do that as the goal of this is to load less on every page.

The quickest short-term fix I can imagine that I'm okay with  is doing this:
```
cp vendor/js/wmd/jquery.wmd.js openlibrary/plugins/openlibrary/js/
```

After that
1) update the import in index.js to this file
2) add
```
window.Attacklab = Attacklab;
```
to the bottom of the copied file.

Thoughts?