@jdlrobson 2019-08-20
Okay this should be working now. Beware Docker filesystem issues (I'm hitting some myself). You may need to cd vendor/js/wmd and do git fetch origin && git checkout origin/master in docker bash if wmd is not working.