@cdrini 2019-08-20
Yeah, that is weird! I think it's because we have a volume for `vendor/`. Looks like if you run `docker-compose exec web git submodule update --recursive` will work, but it won't work if you updated the submodules outside of docker.